Move validate logic into DesktopSnapGroup
